feat: configure widget i18n
