+++ src/constants/locales.ts
@@ -33,7 +33,7 @@ export const SUPPORTED_LOCALES = [
   'zh-CN',
   'zh-TW',
 ] as const
-export type SupportedLocale = typeof SUPPORTED_LOCALES[number]
+export type SupportedLocale = typeof SUPPORTED_LOCALES[number] | 'pseudo'
 
 export const DEFAULT_LOCALE: SupportedLocale = 'en-US'
 
@@ -72,4 +72,5 @@ export const LOCALE_LABEL: { [locale in SupportedLocale]: string } = {
   'vi-VN': 'Tiếng Việt',
   'zh-CN': '简体中文',
   'zh-TW': '繁体中文',
+  pseudo: 'ƥƨèúδô',
 }
